[Release-1.5.0] change getInt->getLong

+++ angel-ps/mllib/src/main/java/com/tencent/angel/ml/task/PredictTask.scala
@@ -28,7 +28,7 @@ import com.tencent.angel.worker.task.{BaseTask, TaskContext}
 
 abstract class PredictTask[KEYIN, VALUEIN](ctx: TaskContext) extends BaseTask[KEYIN, VALUEIN, LabeledData](ctx) {
 
-  val feaNum = conf.getInt(MLConf.ML_FEATURE_NUM, MLConf.DEFAULT_ML_FEATURE_NUM)
+  val feaNum = conf.getLong(MLConf.ML_FEATURE_NUM, MLConf.DEFAULT_ML_FEATURE_NUM)
   val dataFormat = conf.get(MLConf.ML_DATA_FORMAT)
 
   val dataParser = DataParser(dataFormat, feaNum,  negY= true)
